party_app/requirements.txt
Werkzeug==3.1.3
pymongo==4.6.3
python-dotenv==0.19.0
pubnub==9.1.0
Flask==3.1.0
Flask-SocketIO==5.5.0
google-auth==2.36.0
google-auth-oauthlib==1.2.1
Jinja2==3.1.4
oauthlib==3.2.2
requests==2.32.3
requests-oauthlib==2.0.0
yt-dlp==2024.12.23
marshmallow==3.23.2
eventlet
gunicorn
